requirements.txt
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
djoser==2.3.1
pillow==11.2.1
python-decouple==3.8
